+++ tests/can_import_from_another_crate.rs
@@ -15,3 +15,21 @@ fn test_horner() {
     assert_eq!(horner!(x; [2, 3, 4]), 2 + x * (3 + x * 4));
     assert_eq!(horner!(x; [2, 3, 4,]), 2 + x * (3 + x * 4));
 }
+
+#[test]
+fn test_horner_fma() {
+    use polyeval::horner_fma;
+
+    let x = 7;
+
+    assert_eq!(horner_fma!(x; [0]), 0);
+    assert_eq!(horner_fma!(x; [0,]), 0);
+    assert_eq!(horner_fma!(x; 0), 0);
+    assert_eq!(horner_fma!(x; 0,), 0);
+
+    assert_eq!(horner_fma!(x; 2, 3, 4), 2 + x * (3 + x * 4));
+    assert_eq!(horner_fma!(x; 2, 3, 4,), 2 + x * (3 + x * 4));
+
+    assert_eq!(horner_fma!(x; [2, 3, 4]), 2 + x * (3 + x * 4));
+    assert_eq!(horner_fma!(x; [2, 3, 4,]), 2 + x * (3 + x * 4));
+}